requirements-test.txt
build
pytest
pytest-django
pytest-mock
tox
